[cudaaligner] Combine benchmarks into a single binary
Fixes #148
